requirements.txt
Package         Version  
--------------- ---------
backoff         1.8.0    
certifi         2019.6.16
chardet         3.0.4    
ciso8601        2.1.1    
gql             0.1.0    
graphql-core    2.2.1    
graphqlclient   0.2.4    
idna            2.8      
jsonschema      2.6.0    
pip             19.2.1   
pkg-resources   0.0.0    
promise         2.2.1    
python-dateutil 2.8.0    
pytz            2018.4   
requests        2.22.0   
Rx              1.6.1    
setuptools      20.7.0   
simplejson      3.11.1   
singer-python   5.8.0    
six             1.12.0   
typing          3.7.4    
urllib3         1.25.3   
